+++ test/indexer/read_write_test.rb
@@ -87,6 +87,37 @@ describe "Traject::Indexer#process" do
     assert_equal [:one, :two], called, "Both after_processing hooks called, in order"
   end
 
+  it "calls after_processing from #run_after_processing_steps" do
+    @indexer = Traject::Indexer.new(
+      "writer_class_name" => "Traject::NullWriter"
+    )
+    @file = File.open(support_file_path "test_data.utf8.mrc")
+
+    called = []
+
+    @indexer.after_processing do
+      called << :one
+    end
+    @indexer.after_processing do
+      called << :two
+    end
+
+    @indexer.run_after_processing_steps
+    assert_equal [:one, :two], called, "Both after_processing hooks called, in order"
+  end
+
+  it "can't be run twice" do
+    @file = File.open(support_file_path "test_data.utf8.mrc")
+    @indexer = Traject::Indexer.new(
+      "writer_class_name" => "Traject::NullWriter"
+    )
+    @indexer.process(@file)
+
+    assert_raises Traject::Indexer::CompletedStateError do
+      @indexer.process(@file)
+    end
+  end
+
   describe "demo_config.rb" do
     before do
       @indexer = Traject::Indexer.new(